pyproject.toml
[tool.poetry]
name = "nft_project"
version = "0.1.0"
description = ""
authors = ["Seth Mcfeeters <seth.mcfeeters@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^9.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
